requirements.txt
Django==2.2.6
django-cors-headers==3.1.1
django-filter==2.2.0
django-model-utils==3.2.0
djangorestframework==3.10.3
model-mommy==2.0.0
pytz==2019.3
sqlparse==0.3.0
